# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Socious-API
  version: 2.2.6
servers:
  - url: http://localhost:4000
    description: Local server (for development purposes)
  - url: https://devsif.socious.io
    description: dev server (for development purposes)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /projects:
    get:
      summary: get projects
      security:
        - bearerAuth: []
      description: get projects
      tags:
        - Project
      responses:
        '200':
          description: Successful response
      parameters:
        - name: filter.identity
          in: query
          schema:
            type: string
          required: false
    post:
      summary: create projects
      security:
        - bearerAuth: []
      description: create projects
      tags:
        - Project
      parameters:
      - name: Current-Identity
        in: header
        schema:
          type: string
        example: ebd02a72-f530-4d18-9a72-d7aa9ace1271
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  title:
                    type: string
                    example: Title of the Service
                  description:
                    type: string
                    example: Description of the Service
                  status:
                    enum:
                      - DRAFT
                      - EXPIRE
                      - ACTIVE
                  city:
                    type: string
                  country:
                    type: string
                  social_cause:
                    type: string
                  cover_id:
                    type: string
                  wallet_address:
                    type: string
                  wallet_env:
                    type: string
                  website:
                    type: string
                  linkdin:
                    type: string
                  video:
                    type: string
                  problem_statement:
                    type: string
                  solution:
                    type: string
                  goals:
                    type: string
                  total_requested_amount:
                    type: integer
                  cost_beakdown:
                    type: string
                  impact_assessment:
                    type: integer
                  voluntery_contribution:
                    type: string
                  feasibility:
                    type: string
                  categry:
                    enum:
                      - OPEN_INNOVATION
                      - WOMEN_LEADERS
                      - EMERGING_MARKETS
  /projects/{id}:
    parameters:
      - name: id
        required: true
        in: path
        description: Project ID
        schema:
          type: string
    get:
      summary: get project
      security:
        - bearerAuth: []
      description: get project
      tags:
        - Project
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  title:
                    type: string
                    example: Title of the Service
                  description:
                    type: string
                    example: Description of the Service
                  status:
                    enum:
                      - DRAFT
                      - EXPIRE
                      - ACTIVE
                  city:
                    type: string
                  country:
                    type: string
                  social_cause:
                    type: string
                  cover_id:
                    type: string
                  wallet_address:
                    type: string
                  wallet_env:
                    type: string
                  website:
                    type: string
    patch:
      summary: update projects
      security:
        - bearerAuth: []
      description: create projects
      tags:
        - Project
      parameters:
      - name: Current-Identity
        in: header
        schema:
          type: string
        example: ebd02a72-f530-4d18-9a72-d7aa9ace1271
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  title:
                    type: string
                    example: Title of the Service
                  description:
                    type: string
                    example: Description of the Service
                  status:
                    enum:
                      - DRAFT
                      - EXPIRE
                      - ACTIVE
                  city:
                    type: string
                  country:
                    type: string
                  social_cause:
                    type: string
                  cover_id:
                    type: string
                  wallet_address:
                    type: string
                  wallet_env:
                    type: string
                  website:
                    type: string
                  linkdin:
                    type: string
                  video:
                    type: string
                  problem_statement:
                    type: string
                  solution:
                    type: string
                  goals:
                    type: string
                  total_requested_amount:
                    type: integer
                  cost_beakdown:
                    type: string
                  impact_assessment:
                    type: integer
                  voluntery_contribution:
                    type: string
                  feasibility:
                    type: string
                  categry:
                    enum:
                      - OPEN_INNOVATION
                      - WOMEN_LEADERS
                      - EMERGING_MARKETS
    delete:
      summary: delete project
      security:
        - bearerAuth: []
      description: delete project
      tags:
        - Project
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string  
  /projects/{id}/votes:
    post:
      summary: vote for project
      description: vote for project
      security:
        - bearerAuth: []
      tags:
        - Project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
  /projects/{id}/donates:
    post:
      summary: donate to project
      description: donate to project
      security:
        - bearerAuth: []
      tags:
        - Project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the project to donate to
      requestBody:
        content:
          application/json:
            schema:
              properties:
                currency:
                  type: string
                total_amount:
                  type: number
                txid:
                  type: string
                payment_type:
                  type: string
                card_token:
                  type: string
                meta:
                  type: object
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
  /projects/{id}/comments:
    get:
      summary: get all comments
      description: get all comments
      security:
        - bearerAuth: []
      tags:
        - Project
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
    post:
      summary: create comment
      description: create comment
      security:
        - bearerAuth: []
      tags:
        - Project
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              properties:
                content:
                  type: string
                media_id:
                  type: string
                parent_id:
                  type: string
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema: {}
  /projects/comments/{id}:
    get:
      summary: get comment
      description: get comment
      security:
        - bearerAuth: []
      tags:
        - Project
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the comment to retrieve
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: {}
    put:
      summary: update comment
      description: update comment
      security:
        - bearerAuth: []
      tags:
        - Project
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the comment to update
      requestBody:
        content:
          application/json:
            schema:
              properties:
                content:
                  type: string
                media_id:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: {}
    delete:
      summary: delete comment
      description: delete comment
      security:
        - bearerAuth: []
      tags:
        - Project
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the comment to delete
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: {}
  /comments/{id}/likes:
    post:
      summary: like comment
      description: like comment
      security:
        - bearerAuth: []
      tags:
        - Comment
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the comment to like
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema: {}
    delete:
      summary: unlike comment
      description: unlike comment
      security:
        - bearerAuth: []
      tags:
        - Comment
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the comment to unlike
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: {}
  /comments/{id}/reactions:
    post:
      summary: react on comment
      description: react on comment
      security:
        - bearerAuth: []
      tags:
        - Comment
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the comment to react
      requestBody:
        content:
          application/json:
            schema:
              properties:
                reaction:
                  type: string
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema: {}
    delete:
      summary: unreact on comment
      description: unreact on comment
      security:
        - bearerAuth: []
      tags:
        - Comment
      parameters:
        - name: Current-Identity
          in: header
          schema:
            type: string
          required: true
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the comment to unreact
...